Remove others and keep only MPR services
