+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.wso2.gitIssueService</groupId>
-    <artifactId>git-Issue-Service</artifactId>
+    <groupId>org.wso2.mprservice</groupId>
+    <artifactId>MPR-Backend-Service</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>bundle</packaging>
     <parent>
@@ -43,7 +43,7 @@
 
     <properties>
         <export.package>
-            org.wso2.gitissueservice.*,
+            org.wso2.mprservice.*,
         </export.package>
         <import.package>
             *;resolution:=optional
