--- src/main/java/org/wso2/gitissueservice/DependencySummaryProvider.java
@@ -1,50 +1,0 @@
-package org.wso2.gitissueservice;
-
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.utils.URIBuilder;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-
-import java.io.IOException;
-import java.net.URISyntaxException;
-
-/**
- * This class is do the http actions to retrieve the data from ballerina backend for Dependency Dashboard.
- **/
-public class DependencySummaryProvider {
-
-    private static final String HOST_URL = "http://localhost:9091";
-
-    public Object findAllSummaryData() throws IOException, URISyntaxException {
-
-        String response;
-        try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-
-            URIBuilder uriBuilder = new URIBuilder(HOST_URL + "/dependency-data/all");
-            HttpGet httpGet = new HttpGet(uriBuilder.build());
-
-            try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-                response = EntityUtils.toString(response1.getEntity(), "UTF-8");
-            }
-        }
-        return response;
-    }
-
-    public Object findAllProductData() throws IOException, URISyntaxException {
-
-        String response;
-        try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-
-            URIBuilder uriBuilder = new URIBuilder(HOST_URL + "/dependency-data/product-data/all");
-            HttpGet httpGet = new HttpGet(uriBuilder.build());
-
-            try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-                response = EntityUtils.toString(response1.getEntity(), "UTF-8");
-            }
-        }
-        return response;
-    }
-
-}
