--- src/main/java/org/wso2/gitissueservice/utils/DataValueHolder.java
@@ -1,27 +1,0 @@
-package org.wso2.gitissueservice.utils;
-
-import javax.sql.DataSource;
-
-/**
- * This class holding Data source object.
- */
-public class DataValueHolder {
-    private DataSource datasource;
-    private static DataValueHolder dataValueHolder =  new DataValueHolder();
-
-    private DataValueHolder() {
-
-    }
-
-    public static DataValueHolder getInstance() {
-        return dataValueHolder;
-    }
-
-    public void setDataSource(DataSource datasource) {
-        this.datasource = datasource;
-    }
-
-    public DataSource getDataSource() {
-        return datasource;
-    }
-}
