+++ src/main/java/org/wso2/mprservice/MPRServiceProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.wso2.gitissueservice;
+package org.wso2.mprservice;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.CloseableHttpResponse;
@@ -29,20 +29,34 @@ import java.net.URI;
 import java.net.URISyntaxException;
 
 /**
- * This class is do the http actions to retrieve the data from ballerina backend
+ * This class is do the http actions to retrieve the data from ballerina backend for MPR Dashboard.
  **/
-public class GitServiceProvider {
-    private static final String HOST_URL = "http://localhost:9095/gitIssues/";
+public class MPRServiceProvider {
+    private static final String HOST_URL = "http://localhost:9090";
 
 
-    public Object retrieveIssuesFromRepoByLabel(String labels, String repos) throws IOException, URISyntaxException {
+    public Object retrieveProducts() throws IOException, URISyntaxException {
 
         String response;
         try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
 
-            URIBuilder uriBuilder = new URIBuilder(HOST_URL + "repository/label");
-            uriBuilder.addParameter("labels", labels);
-            uriBuilder.addParameter("repos", repos);
+            URIBuilder uriBuilder = new URIBuilder(HOST_URL + "/products");
+            HttpGet httpGet = new HttpGet(uriBuilder.build());
+
+            try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
+                response = EntityUtils.toString(response1.getEntity(), "UTF-8");
+            }
+        }
+        return response;
+    }
+
+    public Object retrieveVersions(String product) throws IOException, URISyntaxException {
+
+        String response;
+        try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
+
+            URIBuilder uriBuilder = new URIBuilder(HOST_URL + "/versions");
+            uriBuilder.addParameter("product", product);
 
             HttpGet httpGet = new HttpGet(uriBuilder.build());
 
@@ -53,33 +67,15 @@ public class GitServiceProvider {
         return response;
     }
 
-    public Object retrieveIssuesFromProduct(String product, String labels) throws IOException, URISyntaxException {
+    public Object retrievePRCountbyStatus(String product, String version) throws IOException, URISyntaxException {
 
         String response;
         try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-            URI uri = new URI(HOST_URL + "product/" + product.replace(" ", "%20"));
-
-            URIBuilder uriBuilder = new URIBuilder(uri);
-            uriBuilder.addParameter("labels", labels);
-
-            HttpGet httpGet = new HttpGet(uriBuilder.build());
-
-            try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-                HttpEntity entity1 = response1.getEntity();
-                response = EntityUtils.toString(entity1, "UTF-8");
-            }
-        }
-        return response;
-    }
-
-    public Object retrieveRepoNamesByProduct(String product) throws IOException, URISyntaxException {
-
-        String response;
-        try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-            URI uri = new URI(HOST_URL + "repos/");
+            URI uri = new URI(HOST_URL + "/prcount");
 
             URIBuilder uriBuilder = new URIBuilder(uri);
             uriBuilder.addParameter("product", product);
+            uriBuilder.addParameter("version", version);
 
             HttpGet httpGet = new HttpGet(uriBuilder.build());
 
@@ -91,13 +87,15 @@ public class GitServiceProvider {
         return response;
     }
 
+    public Object retrieveTotalPRCount(String product, String version) throws IOException, URISyntaxException {
 
-    public Object retrieveProductNames() throws IOException, URISyntaxException {
         String response;
         try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-            URI uri = new URI(HOST_URL + "products/");
+            URI uri = new URI(HOST_URL + "/totalprcount");
 
             URIBuilder uriBuilder = new URIBuilder(uri);
+            uriBuilder.addParameter("product", product);
+            uriBuilder.addParameter("version", version);
 
             HttpGet httpGet = new HttpGet(uriBuilder.build());
